update image posts sous conditions

+++ controllers/post.ctrl.js
@@ -60,8 +60,19 @@ exports.update = async (req, res, next) => {
             const updatedPost = await Post.findOne({
                 where: { id: req.params.id },
             });
-            const filename = updatedPost.image.split("/images/")[1];
-            fs.unlink(`images/${filename}`, async () => {
+            if (updatedPost.image !== null) {
+                const filename = updatedPost.image.split("/images/")[1];
+                fs.unlink(`images/${filename}`, async () => {
+                    updatedPost.text = reqPost.text;
+                    (updatedPost.image = `${req.protocol}://${req.get(
+                        "host"
+                    )}/images/${req.file.filename}`),
+                        await updatedPost
+                            .save()
+                            .then((data) => res.status(200).json({ data }))
+                            .catch((error) => res.status(400).json({ error }));
+                });
+            } else {
                 updatedPost.text = reqPost.text;
                 (updatedPost.image = `${req.protocol}://${req.get(
                     "host"
@@ -70,7 +81,7 @@ exports.update = async (req, res, next) => {
                         .save()
                         .then((data) => res.status(200).json({ data }))
                         .catch((error) => res.status(400).json({ error }));
-            });
+            }
         }
     } else {
         if (reqPost.poster_id === reqPost.user_id || reqPost.admin === true) {
